Fixed all new warnings clang had

+++ include/io.h
@@ -23,7 +23,7 @@ char *IO_get_file_contents(const char *file_path);
  * @param[in] line the line that should be printed.
  * @param[in] offset the offset of token that starts the error.
  */
-void IO_print_error(const char *file_path, int line, int offset);
+void IO_print_error(const char *file_path, long line, long offset);
 
 /**
  * @brief Copy the contents from the file at @p original_file_path to @p
